This should return u32, rename a few vars for clarity

+++ src/register_allocator.rs
@@ -626,8 +626,8 @@ impl IRFunction {
         return self.new_sized_stack_location(tp.size());
     }
 
-    pub fn get_stack_offset_for_location(&self, location: u64, tp: DataType) -> i32 {
-        self.stack_bytes_used as i32 - location as i32 - tp.size() as i32
+    pub fn get_stack_offset_for_location(&self, location: u64, tp: DataType) -> u32 {
+        (self.stack_bytes_used - location as usize - tp.size()) as u32
     }
 
     fn get_index_in_block(&self, block_index: usize, instruction_index: usize) -> Option<usize> {
